# business/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do backend Go — atendimento e automação WhatsApp.",
        "title": "Watink Business API",
        "termsOfService": "https://watink.com.br/terms",
        "contact": {
            "name": "Watink Dev",
            "email": "dev@watink.com.br"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://watink.com.br"
        },
        "version": "2.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Opens a Server-Sent Events stream for real-time updates. Auth via token query param.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "realtime"
                ],
                "summary": "SSE event stream",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Extra rooms (csv): chat:{id}, tickets:{status}, notification, helpdesk-kanban",
                        "name": "rooms",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
...
        "/initial-setup": {
            "post": {
                "description": "Cria o primeiro usuário admin e inicializa o tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setup"
                ],
                "summary": "Configuração inicial do tenant",
                "parameters": [
                    {
                        "description": "Dados de configuração inicial",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token. Formato: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}